// client-id/src/app/core/models/inspection-popup/entrprs-search-model.ts
export class EntrprsSearchModel{
    searchEntrprsNm: string;
    searchBizrNO: string;
    searchRprsntvNm: string;
    searchMnfcturAt: string;
    searchMesurProofIndutyAt: string;
    searchMrnrEmplyrAt: string;
    searchEntrprsSeCode: string;
    searchMrnrKndCode: string;
    searchAtptSeCode: string;
    searchGuGunSeCode: string;
    searchRcognNoNm: string;
    searchMberAt: string;
    searchCategory: string;
    searchType: string;
    searchText: string;
    searchStartDate: string;
    searchEndDate: string;
    startIndex: number;
    endIndex: number;
    constructor(){
        this.searchEntrprsNm = null;
        this.searchBizrNO = null;
        this.searchRprsntvNm = null;
        this.searchMnfcturAt = null;
        this.searchMesurProofIndutyAt = null;
        this.searchMrnrEmplyrAt = null;
        this.searchEntrprsSeCode = null;
        this.searchMrnrKndCode = null;
        this.searchAtptSeCode = null;
        this.searchGuGunSeCode = null;
        this.searchRcognNoNm = null;
        this.searchMberAt = null;
        this.searchCategory = null;
        this.searchType = null;
        this.searchText = null;
        this.searchStartDate = null;
        this.searchEndDate = null;
        this.startIndex = 0;
        this.endIndex = 10;
    }
}